+++ OktaPosh/private/convertToHashTable.ps1
@@ -1,0 +1,10 @@
+function convertToHashTable( [PSCustomObject] $object ) {
+    $ret = @{}
+    if ($object) {
+        foreach ($m in (Get-Member -InputObject $object -MemberType NoteProperty)) {
+            $n = $m.Name
+            $ret[$m.Name] = $object.$n
+        }
+    }
+    $ret
+}