+++ OktaPosh/private/getProp.ps1
@@ -1,0 +1,8 @@
+function getProp( $object, $name, $default = $null )
+{
+    if (Get-Member -InputObject $object -Name $name) {
+        $object.$name
+    } else {
+        $default
+    }
+}